backend/requirements.txt
# FILEPATH: backend/requirements.txt
# @file: Python Dependencies
# @description: The modern async stack for Flodock Platform v2.
# UPDATED: Added 'aiokafka' for the System Nervous System.

# 1. The Spinal Cord (API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 2. The Database (Async ORM)
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.1

# 3. The Brain (Validation & Settings)
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# 4. Security (Auth & Encryption)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2

# 5. Utilities
httpx>=0.26.0
pytz>=2023.3
jmespath>=1.0.1

# 6. Intelligence (AI)
google-genai>=1.0.0

# 7. The Nervous System (Kafka)
aiokafka>=0.10.0


jsonschema>=4.0.0
